+++ .buildkite/steps/copy.sh
@@ -109,6 +109,8 @@ IMAGES=(
 linux_amd64_source_image_id="${1:-}"
 linux_arm64_source_image_id="${1:-}"
 windows_amd64_source_image_id="${2:-}"
+ubuntu2404_amd64_source_image_id="${3:-}"
+ubuntu2404_arm64_source_image_id="${3:-}"
 
 source_region="${AWS_REGION}"
 mapping_file="build/mappings.yml"
@@ -118,6 +120,8 @@ if [ $# -eq 0 ]; then
   linux_amd64_source_image_id=$(buildkite-agent meta-data get "linux_amd64_image_id")
   linux_arm64_source_image_id=$(buildkite-agent meta-data get "linux_arm64_image_id")
   windows_amd64_source_image_id=$(buildkite-agent meta-data get "windows_amd64_image_id")
+  ubuntu2404_amd64_source_image_id=$(buildkite-agent meta-data get "ubuntu2404_amd64_image_id")
+  ubuntu2404_arm64_source_image_id=$(buildkite-agent meta-data get "ubuntu2404_arm64_image_id")
 fi
 
 # If we're not on the main branch or a tag build skip the copy
@@ -127,7 +131,7 @@ if [[ $BUILDKITE_BRANCH != main && $BUILDKITE_TAG != "$BUILDKITE_BRANCH" && ${CO
   cat <<EOF >"$mapping_file"
 Mappings:
   AWSRegion2AMI:
-    ${AWS_REGION} : { linuxamd64: $linux_amd64_source_image_id, linuxarm64: $linux_arm64_source_image_id, windows: $windows_amd64_source_image_id }
+    ${AWS_REGION} : { linuxamd64: $linux_amd64_source_image_id, linuxarm64: $linux_arm64_source_image_id, windows: $windows_amd64_source_image_id, ubuntu2404amd64: $ubuntu2404_amd64_source_image_id, ubuntu2404arm64: $ubuntu2404_arm64_source_image_id }
 EOF
   exit 0
 fi
@@ -137,6 +141,8 @@ if [[ $BUILDKITE_BRANCH == main || $BUILDKITE_TAG == "$BUILDKITE_BRANCH" || ${TA
   tag-ami "$linux_amd64_source_image_id" "$source_region" IsReleased true
   tag-ami "$linux_arm64_source_image_id" "$source_region" IsReleased true
   tag-ami "$windows_amd64_source_image_id" "$source_region" IsReleased true
+  tag-ami "$ubuntu2404_amd64_source_image_id" "$source_region" IsReleased true
+  tag-ami "$ubuntu2404_arm64_source_image_id" "$source_region" IsReleased true
 fi
 
 echo "--- Tagging elastic ci stack release version"
@@ -149,14 +155,18 @@ if [[ $BUILDKITE_TAG == "$BUILDKITE_BRANCH" || ${TAG_VERSION:-false} == true ]];
   tag-ami "$linux_amd64_source_image_id" "$source_region" "Version:${BUILDKITE_TAG}" true
   tag-ami "$linux_arm64_source_image_id" "$source_region" "Version:${BUILDKITE_TAG}" true
   tag-ami "$windows_amd64_source_image_id" "$source_region" "Version:${BUILDKITE_TAG}" true
+  tag-ami "$ubuntu2404_amd64_source_image_id" "$source_region" "Version:${BUILDKITE_TAG}" true
+  tag-ami "$ubuntu2404_arm64_source_image_id" "$source_region" "Version:${BUILDKITE_TAG}" true
 fi
 
 echo "--- Checking if there is a previously copy in the cache bucket"
-s3_mappings_cache=$(printf "s3://%s/mappings-%s-%s-%s-%s.yml" \
+s3_mappings_cache=$(printf "s3://%s/mappings-%s-%s-%s-%s-%s-%s.yml" \
   "${BUILDKITE_AWS_STACK_BUCKET}" \
   "${linux_amd64_source_image_id}" \
   "${linux_arm64_source_image_id}" \
   "${windows_amd64_source_image_id}" \
+  "${ubuntu2404_amd64_source_image_id}" \
+  "${ubuntu2404_arm64_source_image_id}" \
   "${BUILDKITE_BRANCH}")
 
 if aws s3 cp "${s3_mappings_cache}" "$mapping_file"; then
@@ -170,6 +180,8 @@ echo "--- Copying images to other regions"
 linux_amd64_source_image_name=$(get_image_name "$linux_amd64_source_image_id" "$source_region")
 linux_arm64_source_image_name=$(get_image_name "$linux_arm64_source_image_id" "$source_region")
 windows_amd64_source_image_name=$(get_image_name "$windows_amd64_source_image_id" "$source_region")
+ubuntu2404_amd64_source_image_name=$(get_image_name "$ubuntu2404_amd64_source_image_id" "$source_region")
+ubuntu2404_arm64_source_image_name=$(get_image_name "$ubuntu2404_arm64_source_image_id" "$source_region")
 
 # Copy to all other regions
 for region in "${ALL_REGIONS[@]}"; do
@@ -182,8 +194,14 @@ for region in "${ALL_REGIONS[@]}"; do
 
     echo "--- :windows: Copying Windows AMD64 $windows_amd64_source_image_id to $region" >&2
     IMAGES+=("$(copy_ami_to_region "$windows_amd64_source_image_id" "$source_region" "$region" "${windows_amd64_source_image_name}-${region}")")
+
+    echo "--- :ubuntu: Copying Ubuntu AMD64 $ubuntu2404_amd64_source_image_id to $region" >&2
+    IMAGES+=("$(copy_ami_to_region "$ubuntu2404_amd64_source_image_id" "$source_region" "$region" "${ubuntu2404_amd64_source_image_name}-${region}")")
+
+    echo "--- :ubuntu: Copying Ubuntu ARM64 $ubuntu2404_arm64_source_image_id to $region" >&2
+    IMAGES+=("$(copy_ami_to_region "$ubuntu2404_arm64_source_image_id" "$source_region" "$region" "${ubuntu2404_arm64_source_image_name}-${region}")")
   else
-    IMAGES+=("$linux_amd64_source_image_id" "$linux_arm64_source_image_id" "$windows_amd64_source_image_id")
+    IMAGES+=("$linux_amd64_source_image_id" "$linux_arm64_source_image_id" "$windows_amd64_source_image_id" "$ubuntu2404_amd64_source_image_id" "$ubuntu2404_arm64_source_image_id")
   fi
 done
 
@@ -199,6 +217,8 @@ for region in "${ALL_REGIONS[@]}"; do
   linux_amd64_image_id="${IMAGES[0]}"
   linux_arm64_image_id="${IMAGES[1]}"
   windows_amd64_image_id="${IMAGES[2]}"
+  ubuntu2404_amd64_image_id="${IMAGES[3]}"
+  ubuntu2404_arm64_image_id="${IMAGES[4]}"
 
   wait_for_ami_to_be_available "$linux_amd64_image_id" "$region" >&2
 
@@ -224,11 +244,27 @@ for region in "${ALL_REGIONS[@]}"; do
     make_ami_public "$windows_amd64_image_id" "$region"
   fi
 
+  wait_for_ami_to_be_available "$ubuntu2404_amd64_image_id" "$region" >&2
+
+  # Make the ubuntu AMI public if it's not the source image
+  if [[ $ubuntu2404_amd64_image_id != "$ubuntu2404_amd64_source_image_id" ]]; then
+    echo ":ubuntu: Making Ubuntu AMD64 ${ubuntu2404_amd64_image_id} public" >&2
+    make_ami_public "$ubuntu2404_amd64_image_id" "$region"
+  fi
+
+  wait_for_ami_to_be_available "$ubuntu2404_arm64_image_id" "$region" >&2
+
+  # Make the ubuntu ARM AMI public if it's not the source image
+  if [[ $ubuntu2404_arm64_image_id != "$ubuntu2404_arm64_source_image_id" ]]; then
+    echo ":ubuntu: Making Ubuntu ARM64 ${ubuntu2404_arm64_image_id} public" >&2
+    make_ami_public "$ubuntu2404_arm64_image_id" "$region"
+  fi
+
   # Write yaml to file
-  echo "    $region : { linuxamd64: $linux_amd64_image_id, linuxarm64: $linux_arm64_image_id, windows: $windows_amd64_image_id }" >>"$mapping_file"
+  echo "    $region : { linuxamd64: $linux_amd64_image_id, linuxarm64: $linux_arm64_image_id, windows: $windows_amd64_image_id, ubuntu2404amd64: $ubuntu2404_amd64_image_id, ubuntu2404arm64: $ubuntu2404_arm64_image_id }" >>"$mapping_file"
 
   # Shift off the processed images
-  IMAGES=("${IMAGES[@]:3}")
+  IMAGES=("${IMAGES[@]:5}")
 done
 
 echo "--- Uploading mapping to s3 cache"